requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
idna==2.6
Mastodon.py==1.1.1
python-dateutil==2.6.1
pytz==2017.2
requests==2.18.4
six==1.11.0
urllib3==1.22
